Initial transfer of Face to new repository
This is taken from the latest iteration of the Face toolbox and edited for standard content.

Also, all the toolbox docs have been combined into a single set. It would not be too difficult to split them up again, but this way you see both together instead of a lot of individual docs based on 2D or 3D.

Currently I have removed the internal references to other documents, in part as I wasn't sure we should have them.

I did leave the links in the verification list document, though I'm not sure we should (maybe it would be best to just leave the location instead of an actual link).