Proposed update to the Face toolbox
This is based on the planned changes to have the PAD-L2 face toolbox. This removed the full 3D while leaving the wearable mask (that is just taking a printed photo and wearing it like a mask)

This also tries to use the same structure from the fingerprint toolbox update in terms of naming/numbering.